+++ inter_utils.cpp
@@ -20,7 +20,7 @@ Obj str_to_obj(const char *c_str)
   {
     Seq *raw_str = new_seq(n);
     for (int i=0 ; i < n ; i++)
-      raw_str->elems[i] = to_obj(c_str[i]);
+      raw_str->elems[i] = to_obj((long long) c_str[i]);
     raw_str_obj = make_obj(raw_str);
   }
 
@@ -135,7 +135,7 @@ void print(Obj obj)
 {
   generated::Env env;
   memset(&env, 0, sizeof(generated::Env));
-  Obj str_obj = generated::To_Text(obj, to_obj(80), to_obj(0), env);
+  Obj str_obj = generated::To_Text(obj, to_obj(80LL), to_obj(0LL), env);
   int buff_size = char_buffer_size(str_obj);
   char *buffer = (char *) new_obj(nblocks16(buff_size));
   obj_to_str(str_obj, buffer, buff_size);
